Move configuration schema into config package
Relocate configuration_schema to navigate.config and update feature-facing
imports to use the new module path. Keep the old devices module as a
compatibility re-export so existing device imports continue to work unchanged.
