+++ src/navigate/model/devices/configuration_schema.py
@@ -29,148 +29,12 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-# Standard Library Imports
-from dataclasses import dataclass
-from typing import Any, Mapping, Optional, Tuple
+"""Compatibility imports for device configuration schemas."""
 
+from navigate.config.configuration_schema import (
+    CollectionSpec,
+    SettingSpec,
+    merge_configuration_schemas,
+)
 
-@dataclass(frozen=True)
-class SettingSpec:
-    """Description of one user-editable setting or feature parameter.
-
-    ``choices`` defines a closed set of displayed values. ``choice_values`` maps
-    displayed values to saved values when those should differ. ``dynamic_source``
-    identifies choices that should be resolved at render time, and ``depends_on``
-    names the parameter that provides context for dependent choices. ``minimum``,
-    ``exclusive_minimum``, ``maximum``, and ``step`` define a numeric range; they
-    are intentionally mutually exclusive with ``choices``.
-    """
-
-    value_type: type
-    default: Any = None
-    label: Optional[str] = None
-    help_text: Optional[str] = None
-    choices: Optional[Tuple[Any, ...]] = None
-    choice_values: Optional[Mapping[Any, Any]] = None
-    minimum: Optional[float] = None
-    exclusive_minimum: Optional[float] = None
-    maximum: Optional[float] = None
-    step: Optional[float] = None
-    required: bool = False
-    dynamic_source: Optional[str] = None
-    depends_on: Optional[str] = None
-
-    def __post_init__(self) -> None:
-        """Validate a schema definition when its class is imported."""
-        has_range = any(
-            value is not None
-            for value in (
-                self.minimum,
-                self.exclusive_minimum,
-                self.maximum,
-                self.step,
-            )
-        )
-        if self.choices is not None and has_range:
-            raise ValueError(
-                "SettingSpec choices and numeric ranges are mutually exclusive."
-            )
-        if self.choice_values is not None and self.choices is None:
-            raise ValueError("SettingSpec choice_values require choices.")
-        if self.choice_values is not None and any(
-            choice not in self.choice_values for choice in self.choices
-        ):
-            raise ValueError("SettingSpec choice_values must map every choice.")
-        if has_range and self.value_type not in (int, float):
-            raise ValueError("Numeric ranges require an int or float SettingSpec.")
-        if (
-            self.minimum is not None
-            and self.maximum is not None
-            and self.minimum > self.maximum
-        ):
-            raise ValueError("SettingSpec minimum cannot exceed maximum.")
-        if (
-            self.exclusive_minimum is not None
-            and self.maximum is not None
-            and self.exclusive_minimum >= self.maximum
-        ):
-            raise ValueError("SettingSpec exclusive_minimum must be below maximum.")
-        if self.step is not None and self.step <= 0:
-            raise ValueError("SettingSpec step must be greater than zero.")
-
-    @property
-    def display_label(self) -> str:
-        """Return the configured label, falling back to the setting key at render time."""
-        return self.label or ""
-
-
-@dataclass(frozen=True)
-class CollectionSpec:
-    """Description of a repeatable group of persisted configuration values.
-
-    ``storage='mapping'`` stores rows as a YAML mapping, using ``key_field`` as
-    the mapping key and ``value_field`` as its value. ``storage='parallel_mappings'``
-    stores selected row fields as separate YAML mappings keyed by ``key_field``.
-    ``storage='nested_mapping'`` stores calibration rows as
-    ``solvent -> axis -> zoom -> position``. ``storage='single_mapping'``
-    stores one fixed mapping using the item schema keys as fields.
-    ``dynamic_source`` identifies collection fields that should be resolved at
-    render time.
-    """
-
-    item_schema: Mapping[str, SettingSpec]
-    storage: str = "mapping"
-    key_field: str = "name"
-    value_field: str = "value"
-    storage_fields: Optional[Tuple[str, ...]] = None
-    label: Optional[str] = None
-    help_text: Optional[str] = None
-    dynamic_source: Optional[str] = None
-    minimum_items: int = 0
-
-    def __post_init__(self) -> None:
-        """Validate collection metadata at class-definition time."""
-        if self.storage not in {
-            "mapping",
-            "parallel_mappings",
-            "nested_mapping",
-            "single_mapping",
-        }:
-            raise ValueError("CollectionSpec has an unsupported storage type.")
-        if self.storage == "single_mapping":
-            return
-        if self.key_field not in self.item_schema:
-            raise ValueError("CollectionSpec key_field must be in item_schema.")
-        if self.value_field not in self.item_schema:
-            raise ValueError("CollectionSpec value_field must be in item_schema.")
-        if self.storage == "parallel_mappings":
-            if not self.storage_fields or any(
-                field not in self.item_schema for field in self.storage_fields
-            ):
-                raise ValueError(
-                    "Parallel mapping collections require valid storage fields."
-                )
-        if self.storage == "nested_mapping" and set(self.item_schema) != {
-            "solvent",
-            "axis",
-            "zoom",
-            "position",
-        }:
-            raise ValueError(
-                "Nested mapping collections require solvent, axis, zoom, and position fields."
-            )
-        if self.minimum_items < 0:
-            raise ValueError("CollectionSpec minimum_items cannot be negative.")
-
-
-def merge_configuration_schemas(
-    *schemas: Mapping[str, Any],
-) -> dict[str, Any]:
-    """Merge schemas in inheritance order, allowing later schemas to override.
-
-    The caller supplies parent schemas first and child schemas last.
-    """
-    merged: dict[str, Any] = {}
-    for schema in schemas:
-        merged.update(schema)
-    return merged
+__all__ = ["CollectionSpec", "SettingSpec", "merge_configuration_schemas"]
